Add '#' as secret javaagent args delimiter
NOTE: The hash character is the only semi-intuitive character that does not
clash with the regular expression language. It is safe use to on the java
command line because it is specified between quotes after the javaagent equals
sign. This serves as an alternative to the space character as the delimiter for
odd situations.